render: Split off shape/mesh code into its own file in wgpu renderer
