+++ render/wgpu/src/lib.rs
@@ -12,17 +12,21 @@ use ruffle_core::shape_utils::{DrawCommand, DrawPath};
 use std::convert::TryInto;
 use swf::{CharacterId, DefineBitsLossless, Glyph, Shape, Twips};
 
-use crate::pipelines::Pipelines;
 use bytemuck::{Pod, Zeroable};
 use futures::executor::block_on;
 use raw_window_handle::HasRawWindowHandle;
 
+use crate::pipelines::Pipelines;
+use crate::shapes::{Draw, DrawType, GradientUniforms, IncompleteDrawType, Mesh};
+use crate::utils::create_buffer_with_data;
+
 type Error = Box<dyn std::error::Error>;
 
 #[macro_use]
-mod macros;
+mod utils;
 
 mod pipelines;
+mod shapes;
 
 pub struct WGPURenderBackend {
     window_surface: wgpu::Surface,
@@ -87,21 +91,6 @@ struct GPUVertex {
 
 unsafe impl Pod for GPUVertex {}
 unsafe impl Zeroable for GPUVertex {}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug)]
-struct GradientUniforms {
-    gradient_type: i32,
-    num_colors: u32,
-    repeat_mode: i32,
-    focal_point: f32,
-    // TODO: pack this more efficiently. Alignment forces a float[16] to be aligned as a float4[16].
-    ratios: [[f32; 4]; 16],
-    colors: [[f32; 4]; 16],
-}
-
-unsafe impl Pod for GradientUniforms {}
-unsafe impl Zeroable for GradientUniforms {}
 
 impl WGPURenderBackend {
     pub fn new<W: HasRawWindowHandle>(window: &W, size: (u32, u32)) -> Result<Self, Error> {
@@ -1420,279 +1409,6 @@ struct Texture {
     texture: wgpu::Texture,
 }
 
-#[derive(Debug)]
-struct Mesh {
-    draws: Vec<Draw>,
-    transforms: wgpu::Buffer,
-    colors: wgpu::Buffer,
-    shape_id: CharacterId,
-}
-
-#[derive(Debug)]
-struct Draw {
-    draw_type: DrawType,
-    vertex_buffer: wgpu::Buffer,
-    index_buffer: wgpu::Buffer,
-    bind_group: wgpu::BindGroup,
-    index_count: u32,
-}
-
-#[derive(Debug)]
-enum DrawType {
-    Color,
-    Gradient {
-        texture_transforms: wgpu::Buffer,
-        gradient: wgpu::Buffer,
-    },
-    Bitmap {
-        texture_transforms: wgpu::Buffer,
-        texture_view: wgpu::TextureView,
-        id: CharacterId,
-    },
-}
-
-#[derive(Debug)]
-#[allow(clippy::large_enum_variant)]
-enum IncompleteDrawType {
-    Color,
-    Gradient {
-        texture_transform: [[f32; 4]; 4],
-        gradient: GradientUniforms,
-    },
-    Bitmap {
-        texture_transform: [[f32; 4]; 4],
-        is_smoothed: bool,
-        is_repeating: bool,
-        texture_view: wgpu::TextureView,
-        id: CharacterId,
-    },
-}
-
-impl IncompleteDrawType {
-    pub fn name(&self) -> &'static str {
-        match self {
-            IncompleteDrawType::Color => "Color",
-            IncompleteDrawType::Gradient { .. } => "Gradient",
-            IncompleteDrawType::Bitmap { .. } => "Bitmap",
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub fn build(
-        self,
-        device: &wgpu::Device,
-        transforms_ubo: &wgpu::Buffer,
-        colors_ubo: &wgpu::Buffer,
-        vertex_buffer: wgpu::Buffer,
-        index_buffer: wgpu::Buffer,
-        index_count: u32,
-        pipelines: &Pipelines,
-        shape_id: CharacterId,
-        draw_id: usize,
-    ) -> Draw {
-        match self {
-            IncompleteDrawType::Color => {
-                let bind_group_label =
-                    create_debug_label!("Shape {} (color) draw {} bindgroup", shape_id, draw_id);
-                let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-                    layout: &pipelines.color.bind_layout,
-                    bindings: &[
-                        wgpu::Binding {
-                            binding: 0,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: transforms_ubo,
-                                range: 0..std::mem::size_of::<Transforms>() as u64,
-                            },
-                        },
-                        wgpu::Binding {
-                            binding: 1,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: colors_ubo,
-                                range: 0..std::mem::size_of::<ColorAdjustments>() as u64,
-                            },
-                        },
-                    ],
-                    label: bind_group_label.as_deref(),
-                });
-
-                Draw {
-                    draw_type: DrawType::Color,
-                    vertex_buffer,
-                    index_buffer,
-                    bind_group,
-                    index_count,
-                }
-            }
-            IncompleteDrawType::Gradient {
-                texture_transform,
-                gradient,
-            } => {
-                let tex_transforms_ubo = create_buffer_with_data(
-                    device,
-                    bytemuck::cast_slice(&[texture_transform]),
-                    wgpu::BufferUsage::UNIFORM,
-                    create_debug_label!(
-                        "Shape {} draw {} textransforms ubo transfer buffer",
-                        shape_id,
-                        draw_id
-                    ),
-                );
-
-                let gradient_ubo = create_buffer_with_data(
-                    device,
-                    bytemuck::cast_slice(&[gradient]),
-                    wgpu::BufferUsage::UNIFORM,
-                    create_debug_label!(
-                        "Shape {} draw {} gradient ubo transfer buffer",
-                        shape_id,
-                        draw_id
-                    ),
-                );
-
-                let bind_group_label =
-                    create_debug_label!("Shape {} (gradient) draw {} bindgroup", shape_id, draw_id);
-                let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-                    layout: &pipelines.gradient.bind_layout,
-                    bindings: &[
-                        wgpu::Binding {
-                            binding: 0,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: transforms_ubo,
-                                range: 0..std::mem::size_of::<Transforms>() as u64,
-                            },
-                        },
-                        wgpu::Binding {
-                            binding: 1,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: &tex_transforms_ubo,
-                                range: 0..std::mem::size_of::<TextureTransforms>() as u64,
-                            },
-                        },
-                        wgpu::Binding {
-                            binding: 2,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: colors_ubo,
-                                range: 0..std::mem::size_of::<ColorAdjustments>() as u64,
-                            },
-                        },
-                        wgpu::Binding {
-                            binding: 3,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: &gradient_ubo,
-                                range: 0..std::mem::size_of::<GradientUniforms>() as u64,
-                            },
-                        },
-                    ],
-                    label: bind_group_label.as_deref(),
-                });
-
-                Draw {
-                    draw_type: DrawType::Gradient {
-                        texture_transforms: tex_transforms_ubo,
-                        gradient: gradient_ubo,
-                    },
-                    vertex_buffer,
-                    index_buffer,
-                    bind_group,
-                    index_count,
-                }
-            }
-            IncompleteDrawType::Bitmap {
-                texture_transform,
-                is_smoothed,
-                is_repeating,
-                texture_view,
-                id,
-            } => {
-                let tex_transforms_ubo = create_buffer_with_data(
-                    device,
-                    bytemuck::cast_slice(&[texture_transform]),
-                    wgpu::BufferUsage::UNIFORM,
-                    create_debug_label!(
-                        "Shape {} draw {} textransforms ubo transfer buffer",
-                        shape_id,
-                        draw_id
-                    ),
-                );
-
-                let address_mode = if is_repeating {
-                    wgpu::AddressMode::Repeat
-                } else {
-                    wgpu::AddressMode::ClampToEdge
-                };
-
-                let filter = if is_smoothed {
-                    wgpu::FilterMode::Linear
-                } else {
-                    wgpu::FilterMode::Nearest
-                };
-
-                let sampler = device.create_sampler(&wgpu::SamplerDescriptor {
-                    address_mode_u: address_mode,
-                    address_mode_v: address_mode,
-                    address_mode_w: address_mode,
-                    mag_filter: filter,
-                    min_filter: filter,
-                    mipmap_filter: filter,
-                    lod_min_clamp: 0.0,
-                    lod_max_clamp: 100.0,
-                    compare: wgpu::CompareFunction::Undefined,
-                });
-
-                let bind_group_label =
-                    create_debug_label!("Shape {} (bitmap) draw {} bindgroup", shape_id, draw_id);
-                let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-                    layout: &pipelines.bitmap.bind_layout,
-                    bindings: &[
-                        wgpu::Binding {
-                            binding: 0,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: transforms_ubo,
-                                range: 0..std::mem::size_of::<Transforms>() as u64,
-                            },
-                        },
-                        wgpu::Binding {
-                            binding: 1,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: &tex_transforms_ubo,
-                                range: 0..std::mem::size_of::<TextureTransforms>() as u64,
-                            },
-                        },
-                        wgpu::Binding {
-                            binding: 2,
-                            resource: wgpu::BindingResource::Buffer {
-                                buffer: colors_ubo,
-                                range: 0..std::mem::size_of::<ColorAdjustments>() as u64,
-                            },
-                        },
-                        wgpu::Binding {
-                            binding: 3,
-                            resource: wgpu::BindingResource::TextureView(&texture_view),
-                        },
-                        wgpu::Binding {
-                            binding: 4,
-                            resource: wgpu::BindingResource::Sampler(&sampler),
-                        },
-                    ],
-                    label: bind_group_label.as_deref(),
-                });
-
-                Draw {
-                    draw_type: DrawType::Bitmap {
-                        texture_transforms: tex_transforms_ubo,
-                        texture_view,
-                        id,
-                    },
-                    vertex_buffer,
-                    index_buffer,
-                    bind_group,
-                    index_count,
-                }
-            }
-        }
-    }
-}
-
 #[allow(clippy::many_single_char_names)]
 fn swf_to_gl_matrix(m: swf::Matrix) -> [[f32; 4]; 4] {
     let tx = m.translate_x.get() as f32;
@@ -1773,19 +1489,4 @@ impl StrokeVertexConstructor<GPUVertex> for RuffleVertexCtor {
             color: self.color,
         }
     }
-}
-
-fn create_buffer_with_data(
-    device: &wgpu::Device,
-    data: &[u8],
-    usage: wgpu::BufferUsage,
-    label: Option<String>,
-) -> wgpu::Buffer {
-    let mapped = device.create_buffer_mapped(&wgpu::BufferDescriptor {
-        size: data.len() as u64,
-        usage,
-        label: label.as_deref(),
-    });
-    mapped.data.copy_from_slice(data);
-    mapped.finish()
 }
