--- render/wgpu/src/macros.rs
@@ -1,9 +1,0 @@
-macro_rules! create_debug_label {
-    ($($arg:tt)*) => (
-        if cfg!(feature = "render_debug_labels") {
-            Some(format!($($arg)*))
-        } else {
-            None
-        }
-    )
-}
